Revert "Merge branch 'main' into current-playlist-page-2"
This reverts commit b8d543510a63e23e8f1692310f45ea9eb2db79bc, reversing
changes made to 202640167bcb8d8dc1cbec9ba221a119b3fa80c1.
