+++ client/src/pages/search/components/SongResult.js
@@ -23,11 +23,11 @@ const SongResult = ({ className, thumbnailUrl, songName, artistName, artists, du
   useEffect(() => {
     if (songObject) {
       const parsed = {
-        name: songObject.name,
-        artist: songObject.artist,
-        type: songObject.type,
-        link: songObject.link,
-        imageLink: songObject.imageLink,
+        name: songObject?.songName,
+        artist: songObject?.artists[0],
+        type: 'spotify',
+        link: songObject?.songLink,
+        imageLink: songObject?.thumbnailUrl
       };
       setparsedSongObject(parsed);
     }
@@ -92,9 +92,7 @@ const SongResult = ({ className, thumbnailUrl, songName, artistName, artists, du
         </div>
         <div className="details">
           <div className="name">{songName}</div>
-
-          <div className="artist">{artistName}</div>
-
+          <div className="artist">{artists}</div>
           <div className="artist-name">{artistName}</div>
         </div>
       </div>
