+++ client/src/pages/search/redux/actionTypes.js
@@ -1,7 +1,10 @@
 export const actionTypes = {
-    GET_NEXT_SPOTIFY: 'spotify/getNext',
-    GET_NEXT_YOUTUBE: 'youtube/getNextVideos',
-    GET_SPOTIFY: 'spotify/get',  
-    GET_YOUTUBE: 'youtube/get',
+    GET_SPOTIFY_TRACKS: 'spotify/getTracks',
+    GET_SPOTIFY_ALBUMS: 'spotify/getAlbums',
+    GET_SPOTIFY_PLAYLISTS: 'spotify/getPlaylists',
+    GET_YOUTUBE_VIDEOS: 'youtube/getVideos',
+    GET_YOUTUBE_PLAYLISTS: 'youtube/getPlaylists',
+    GET_SPOTIFY: 'spotify/get', // temp
+    GET_YOUTUBE: 'youtube/get', // temp
     GET_UNIFI: 'unifi/get',
 }
