+++ client/src/pages/search/redux/reducer.js
@@ -1,6 +1,6 @@
 import { createSlice } from '@reduxjs/toolkit';
 import { REQUEST_STATE } from './utils';
-import { getNextSpotifyAsync, getNextYoutubeAsync, getSpotifyAsync, getYoutubeAsync } from './thunks';
+import { getSpotifyAsync } from './thunks';
 
 
 // because we get rate limited and have limited search results, we need to query for all possible data
@@ -14,20 +14,22 @@ const INITIAL_STATE = {
         'albumsNext': null,
         'playlistsNext': null,
     }, // TODO add artists
-    youtube: { 
-        'videos': [],
-        'playlists': [],
-        'videosNext': null,
-        'playlistsNext': null,
-    },
+    youtube: { videos: [], playlists: [] }, // TODO add channels
     unifi: { playlists: [] }, // TODO: add users
     getSpotify: REQUEST_STATE.IDLE,
-    getSpotifyNext: REQUEST_STATE.IDLE,
     getYoutube: REQUEST_STATE.IDLE,
-    getYoutubeNext: REQUEST_STATE.IDLE,
     getUnifi: REQUEST_STATE.IDLE,
     errors: null
 };
+
+
+
+
+
+// https://developer.spotify.com/documentation/web-api/reference/get-recommendations
+// -> good initial query for recommendations on home page
+
+
 
 
 const searchSlice = createSlice({
@@ -40,84 +42,39 @@ const searchSlice = createSlice({
         });
         builder.addCase(getSpotifyAsync.fulfilled, (state, action) => {
             state.getSpotify = REQUEST_STATE.FULFILLED;
-
-            if ('tracks' in action.payload) {
-                state.spotify.tracks = action.payload.tracks.items;
-                state.spotify.tracksNext = action.payload.tracks.next;
-            }
+            state.spotify.tracks = action.payload.tracks.items;
+            state.spotify.tracksNext = action.payload.tracks.next;
+            // state.nextPageLink = action.payload.tracks.next;
             
-            if ('albums' in action.payload) {
-                state.spotify.albums=action.payload.albums.items;
-                state.spotify.albumsNext = action.payload.albums.next;
-            }
+            // console.log(action.payload.albums)
+            state.spotify.albums=action.payload.albums.items;
+            state.spotify.albumsNext = action.payload.albums.next;
+            // state.spotify.albums = parseSpotifyAlbums(action.payload.albums.items);
+            // console.log(JSON.stringify(state.spotify.albums[0]));
 
-            if ('playlists' in action.payload) {
-                state.spotify.playlists = action.payload.playlists.items;
-                state.spotify.playlistsNext = action.payload.playlists.next;
-            }
+            state.spotify.playlists = action.payload.playlists.items;
+            state.spotify.playlistsNext = action.payload.playlists.next;
+            
+            // state.spotify.artists = action.payload.artists;
+            // state.getSpotify.playlists = action.payload.playlists;
         });
         builder.addCase(getSpotifyAsync.rejected, (state, action) => {
             state.getSpotify = REQUEST_STATE.REJECTED;
             state.errors = action.payload;
         });
-        builder.addCase(getNextSpotifyAsync.pending, (state, action) => {
-            state.getSpotifyNext = REQUEST_STATE.PENDING;
-        });
-        builder.addCase(getNextSpotifyAsync.fulfilled, (state, action) => {
-            state.getSpotifyNext = REQUEST_STATE.FULFILLED;
-
-            if ('tracks' in action.payload) {
-                state.spotify.tracks = [...state.spotify.tracks, ...action.payload.tracks.items];
-                state.spotify.tracksNext = action.payload.tracks.next;
-            }
-
-            if ('albums' in action.payload) {
-                state.spotify.albums = [...state.spotify.albums, ...action.payload.albums.items];
-                state.spotify.albumsNext = action.payload.albums.next;
-            }
-
-            if ('playlists' in action.payload) {
-                state.spotify.playlists = [...state.spotify.playlists, ...action.payload.playlists.items];
-                state.spotify.playlistsNext = action.payload.playlists.next;
-            }
-
-        });
-        builder.addCase(getNextSpotifyAsync.rejected, (state, action) => {
-            state.getSpotifyNext = REQUEST_STATE.REJECTED;
-            state.errors = action.payload;
-        });
-        builder.addCase(getYoutubeAsync.pending, (state, action) => {
-            state.getYoutube = REQUEST_STATE.PENDING;
-        });
-        builder.addCase(getYoutubeAsync.fulfilled, (state, action) => {
-            state.getYoutube = REQUEST_STATE.FULFILLED;
-            if ('videos' in action.payload) {
-                state.youtube.videos = action.payload.videos.items;
-                state.youtube.videosNext = action.payload.videos.next;
-            }
-            if ('playlists' in action.payload) {
-                state.youtube.playlists = action.payload.playlists.items;
-                state.youtube.playlistsNext = action.payload.playlists.next;
-            }
-        });
-        builder.addCase(getYoutubeAsync.rejected, (state, action) => {
-            state.getYoutube = REQUEST_STATE.REJECTED;
-            state.errors = action.payload;
-        });
-        builder.addCase(getNextYoutubeAsync.pending, (state, action) => {
-            state.getYoutubeNext = REQUEST_STATE.PENDING;
-        });
-        builder.addCase(getNextYoutubeAsync.fulfilled, (state, action) => {
-            state.getYoutubeNext = REQUEST_STATE.FULFILLED;
-            if ('videos' in action.payload) {
-                state.youtube.videos = [...state.youtube.videos, ...action.payload.videos.items];
-                state.youtube.videosNext = action.payload.videos.next;
-            }
-            if ('playlists' in action.payload) {
-                state.youtube.playlists = [...state.youtube.playlists, ...action.payload.playlists.items];
-                state.youtube.playlistsNext = action.payload.playlists.next;
-            }
-        });
+        // builder.addCase(getYoutubeAsync.pending, (state, action) => {
+        //     state.getYoutube = REQUEST_STATE.PENDING;
+        // });
+        // builder.addCase(getYoutubeAsync.fulfilled, (state, action) => {
+        //     state.getYoutube = REQUEST_STATE.FULFILLED;
+        //     state.youtubeVideos = action.payload.videos;
+        //     state.youtubePlaylists = action.payload.playlists;
+        //     state.youtubeChannels = action.payload.channels;
+        // });
+        // builder.addCase(getYoutubeAsync.rejected, (state, action) => {
+        //     state.getYoutube = REQUEST_STATE.REJECTED;
+        //     state.errors = action.payload;
+        // });
         // builder.addCase(getUnifiAsync.pending, (state, action) => {
         //     state.getUnifi = REQUEST_STATE.PENDING;
         // });
