+++ client/src/pages/search/samplesearches/drake.json
@@ -1,0 +1,191 @@
+{
+    "artists": {
+      "href": "https://api.spotify.com/v1/search?query=drake&type=artist&market=CA&locale=en-US%2Cen%3Bq%3D0.9&offset=0&limit=5",
+      "items": [
+        {
+          "external_urls": {
+            "spotify": "https://open.spotify.com/artist/3TVXtAsR1Inumwj472S9r4"
+          },
+          "followers": {
+            "href": null,
+            "total": 76770743
+          },
+          "genres": [
+            "canadian hip hop",
+            "canadian pop",
+            "hip hop",
+            "rap",
+            "toronto rap"
+          ],
+          "href": "https://api.spotify.com/v1/artists/3TVXtAsR1Inumwj472S9r4",
+          "id": "3TVXtAsR1Inumwj472S9r4",
+          "images": [
+            {
+              "height": 640,
+              "url": "https://i.scdn.co/image/ab6761610000e5eb4293385d324db8558179afd9",
+              "width": 640
+            },
+            {
+              "height": 320,
+              "url": "https://i.scdn.co/image/ab676161000051744293385d324db8558179afd9",
+              "width": 320
+            },
+            {
+              "height": 160,
+              "url": "https://i.scdn.co/image/ab6761610000f1784293385d324db8558179afd9",
+              "width": 160
+            }
+          ],
+          "name": "Drake",
+          "popularity": 95,
+          "type": "artist",
+          "uri": "spotify:artist:3TVXtAsR1Inumwj472S9r4"
+        },
+        {
+          "external_urls": {
+            "spotify": "https://open.spotify.com/artist/4O15NlyKLIASxsJ0PrXPfz"
+          },
+          "followers": {
+            "href": null,
+            "total": 15337406
+          },
+          "genres": [
+            "melodic rap",
+            "philly rap",
+            "rage rap",
+            "rap",
+            "trap"
+          ],
+          "href": "https://api.spotify.com/v1/artists/4O15NlyKLIASxsJ0PrXPfz",
+          "id": "4O15NlyKLIASxsJ0PrXPfz",
+          "images": [
+            {
+              "height": 640,
+              "url": "https://i.scdn.co/image/ab6761610000e5eb1234d2f516796badbdf16a89",
+              "width": 640
+            },
+            {
+              "height": 320,
+              "url": "https://i.scdn.co/image/ab676161000051741234d2f516796badbdf16a89",
+              "width": 320
+            },
+            {
+              "height": 160,
+              "url": "https://i.scdn.co/image/ab6761610000f1781234d2f516796badbdf16a89",
+              "width": 160
+            }
+          ],
+          "name": "Lil Uzi Vert",
+          "popularity": 86,
+          "type": "artist",
+          "uri": "spotify:artist:4O15NlyKLIASxsJ0PrXPfz"
+        },
+        {
+          "external_urls": {
+            "spotify": "https://open.spotify.com/artist/4mskfuiHWZ3nX3qAdeaGiR"
+          },
+          "followers": {
+            "href": null,
+            "total": 116740
+          },
+          "genres": [],
+          "href": "https://api.spotify.com/v1/artists/4mskfuiHWZ3nX3qAdeaGiR",
+          "id": "4mskfuiHWZ3nX3qAdeaGiR",
+          "images": [
+            {
+              "height": 640,
+              "url": "https://i.scdn.co/image/ab6761610000e5eb2061bbe8748cb2ec6b72ca52",
+              "width": 640
+            },
+            {
+              "height": 320,
+              "url": "https://i.scdn.co/image/ab676161000051742061bbe8748cb2ec6b72ca52",
+              "width": 320
+            },
+            {
+              "height": 160,
+              "url": "https://i.scdn.co/image/ab6761610000f1782061bbe8748cb2ec6b72ca52",
+              "width": 160
+            }
+          ],
+          "name": "Drake Milligan",
+          "popularity": 53,
+          "type": "artist",
+          "uri": "spotify:artist:4mskfuiHWZ3nX3qAdeaGiR"
+        },
+        {
+          "external_urls": {
+            "spotify": "https://open.spotify.com/artist/29ijED2bnnprp2TciAK1aO"
+          },
+          "followers": {
+            "href": null,
+            "total": 109464
+          },
+          "genres": [],
+          "href": "https://api.spotify.com/v1/artists/29ijED2bnnprp2TciAK1aO",
+          "id": "29ijED2bnnprp2TciAK1aO",
+          "images": [
+            {
+              "height": 640,
+              "url": "https://i.scdn.co/image/ab6761610000e5ebfcdf7ed1c10fd9f03bd2e14e",
+              "width": 640
+            },
+            {
+              "height": 320,
+              "url": "https://i.scdn.co/image/ab67616100005174fcdf7ed1c10fd9f03bd2e14e",
+              "width": 320
+            },
+            {
+              "height": 160,
+              "url": "https://i.scdn.co/image/ab6761610000f178fcdf7ed1c10fd9f03bd2e14e",
+              "width": 160
+            }
+          ],
+          "name": "Drake White",
+          "popularity": 52,
+          "type": "artist",
+          "uri": "spotify:artist:29ijED2bnnprp2TciAK1aO"
+        },
+        {
+          "external_urls": {
+            "spotify": "https://open.spotify.com/artist/0p4ViyfJUTW0IT4SCBLexf"
+          },
+          "followers": {
+            "href": null,
+            "total": 318838
+          },
+          "genres": [
+            "cali rap"
+          ],
+          "href": "https://api.spotify.com/v1/artists/0p4ViyfJUTW0IT4SCBLexf",
+          "id": "0p4ViyfJUTW0IT4SCBLexf",
+          "images": [
+            {
+              "height": 640,
+              "url": "https://i.scdn.co/image/ab6761610000e5eba603aef1d62687aa6af6f01f",
+              "width": 640
+            },
+            {
+              "height": 320,
+              "url": "https://i.scdn.co/image/ab67616100005174a603aef1d62687aa6af6f01f",
+              "width": 320
+            },
+            {
+              "height": 160,
+              "url": "https://i.scdn.co/image/ab6761610000f178a603aef1d62687aa6af6f01f",
+              "width": 160
+            }
+          ],
+          "name": "Drakeo the Ruler",
+          "popularity": 61,
+          "type": "artist",
+          "uri": "spotify:artist:0p4ViyfJUTW0IT4SCBLexf"
+        }
+      ],
+      "limit": 5,
+      "next": "https://api.spotify.com/v1/search?query=drake&type=artist&market=CA&locale=en-US%2Cen%3Bq%3D0.9&offset=5&limit=5",
+      "offset": 0,
+      "previous": null,
+      "total": 8
+    }
+  }
