+++ client/src/pages/search/samplesearches/sampleaudiofeatures.json
@@ -1,0 +1,46 @@
+// curl --request GET \
+//   --url 'https://api.spotify.com/v1/audio-features?ids=4bjN59DRXFRxBE1g5ne6B1%2C4eO8lE5u4EONb3cIREyMDv' \
+//   --header 'Authorization: Bearer 1POdFZRZbvb...qqillRxMr2z'
+
+
+{
+    "audio_features": [
+      {
+        "danceability": 0.697,
+        "energy": 0.837,
+        "key": 7,
+        "loudness": -2.651,
+        "mode": 1,
+        "speechiness": 0.304,
+        "acousticness": 0.313,
+        "instrumentalness": 0,
+        "liveness": 0.471,
+        "valence": 0.831,
+        "tempo": 85.097,
+        "time_signature": 5,
+        "analysis_url": "https://api.spotify.com/v1/audio-analysis/4bjN59DRXFRxBE1g5ne6B1",
+        
+        
+      },
+      {
+        "danceability": 0.4,
+        "energy": 0.187,
+        "key": 7,
+        "loudness": -11.204,
+        "mode": 1,
+        "speechiness": 0.0382,
+        "acousticness": 0.926,
+        "instrumentalness": 0.0000149,
+        "liveness": 0.0906,
+        "valence": 0.16,
+        "tempo": 122.029,
+        "type": "audio_features",
+        "id": "4eO8lE5u4EONb3cIREyMDv",
+        "uri": "spotify:track:4eO8lE5u4EONb3cIREyMDv",
+        "track_href": "https://api.spotify.com/v1/tracks/4eO8lE5u4EONb3cIREyMDv",
+        "analysis_url": "https://api.spotify.com/v1/audio-analysis/4eO8lE5u4EONb3cIREyMDv",
+        "duration_ms": 277253,
+        "time_signature": 4
+      }
+    ]
+  }
