--- client/src/pages/search/samplesearches/sp-search-album.txt
@@ -1,5 +1,0 @@
-original:
-localhost:3001/sp-search?q=haegeum&type=album&accessToken=BQBs2kREPHpnNir26VTOs8pN_EPMep-FjEZbRLH6b-xyoZAVZy89q_ugtsbn6UwItQoX3Dc8bjdl_TsAKlHTPBcLYm0LMCmUYIGBZpkcTlo2aWeNiAHb8gAAmkHg3hHTz91Iz0XUDc9tinlrZcNhaPs4DEueQejD2ZtJk0V-ZOlh6Y8_6Sfl7aT0fYd0GLaR7zPbbPbQ0asUqEb0af2Pz1RBdbNGmPI1HrH5wfZm
-
-next:
-localhost:3001/sp-search/next?cookieId=60ef403e-0709-4bd3-b9c0-3514d607dbcc&type=album&accessToken=BQBs2kREPHpnNir26VTOs8pN_EPMep-FjEZbRLH6b-xyoZAVZy89q_ugtsbn6UwItQoX3Dc8bjdl_TsAKlHTPBcLYm0LMCmUYIGBZpkcTlo2aWeNiAHb8gAAmkHg3hHTz91Iz0XUDc9tinlrZcNhaPs4DEueQejD2ZtJk0V-ZOlh6Y8_6Sfl7aT0fYd0GLaR7zPbbPbQ0asUqEb0af2Pz1RBdbNGmPI1HrH5wfZm
