--- client/src/pages/search/samplesearches/sp-search-all.txt
@@ -1,1 +1,0 @@
-localhost:3001/sp-search?q=beatles&accessToken=BQBs2kREPHpnNir26VTOs8pN_EPMep-FjEZbRLH6b-xyoZAVZy89q_ugtsbn6UwItQoX3Dc8bjdl_TsAKlHTPBcLYm0LMCmUYIGBZpkcTlo2aWeNiAHb8gAAmkHg3hHTz91Iz0XUDc9tinlrZcNhaPs4DEueQejD2ZtJk0V-ZOlh6Y8_6Sfl7aT0fYd0GLaR7zPbbPbQ0asUqEb0af2Pz1RBdbNGmPI1HrH5wfZm
