--- client/src/pages/search/samplesearches/yt-search-all.txt
@@ -1,1 +1,0 @@
-localhost:3001/yt-search?q=overwatch
