--- client/src/pages/search/samplesearches/yt-search-playlists.txt
@@ -1,1 +1,0 @@
-localhost:3001/yt-search/playlists/PLOHoVaTp8R7dfrJW5pumS0iD_dhlXKv17
