+++ client/src/pages/search/searchPage.js
@@ -10,7 +10,7 @@ import PlaylistResult from './components/PlaylistResult';
 import Filters from './components/Filters';
 import PlaylistCreator from '../playlists/components/PlaylistCreator';
 import { useSelector, useDispatch } from 'react-redux';
-import { getSpotifyAsync, getYoutubeAsync } from './redux/thunks';
+import { getSpotifyAsync } from './redux/thunks';
 import debounce from 'lodash.debounce';
 // import ScrollingComponent from './ScrollingComponent';
 
@@ -62,8 +62,7 @@ const SearchPage = () => {
   useEffect(() => {
     console.log("calling it with " + searchTerm)
     if (searchTerm === '') return; // make it load recommended songs from spotify
-    dispatch(getSpotifyAsync({accessToken:accessToken, query:searchTerm}));
-    dispatch(getYoutubeAsync({query:searchTerm}));
+    dispatch(getSpotifyAsync({ accessToken: accessToken, query: searchTerm }));
   }, [searchTerm]);
 
 
@@ -119,6 +118,42 @@ const SearchPage = () => {
   }
 
 
+
+  const spotifySongList = [
+    {
+      thumbnailUrl: 'https://i.scdn.co/image/ab67616d0000b273fa9247b68471b82d2125651e',
+      songName: 'Haegeum',
+      artistName: 'Agust D',
+      views: '123456',
+      duration: '1:53',
+      songLink: 'https://open.spotify.com/track/4bjN59DRXFRxBE1g5ne6B1?si=d40939f4f897437d'
+    },
+    {
+      thumbnailUrl: 'https://i1.sndcdn.com/artworks-Vi8kWdDLyiHb-0-t500x500.jpg',
+      songName: 'One Dance',
+      artistName: 'Drake',
+      views: '6942000',
+      duration: '2:53',
+      songLink: 'https://open.spotify.com/track/1zi7xx7UVEFkmKfv06H8x0?si=d0226490edb4470b'
+    },
+    {
+      thumbnailUrl: 'https://static.wikia.nocookie.net/the-bangtan-boys/images/f/fb/Love_Yourself_%27Tear%27_album_cover.jpg',
+      songName: 'FAKE LOVE',
+      artistName: 'BTS',
+      views: '34100000',
+      duration: '4:02',
+      songLink: 'https://open.spotify.com/track/6m1TWFMeon7ai9XLOzdbiR?si=85e1028488b9479c'
+    },
+    {
+      thumbnailUrl: 'https://i.ytimg.com/vi/MAihF124EbE/maxresdefault.jpg',
+      songName: 'Sugar',
+      artistName: 'Maroon 5',
+      views: '20001000',
+      duration: '3:55',
+      songLink: 'https://open.spotify.com/track/2iuZJX9X9P0GKaE93xcPjk?si=3a866fb1a7ec4062'
+    }
+  ]
+
   return (
     <div className='search-page'>
       <SearchBar placeholder='Search for songs, albums, artists...' searchCallback={(input) => { performSearch(input) }} />
@@ -134,13 +169,13 @@ const SearchPage = () => {
         {spotifyTracks.map((song) => (
           <SongResult
             className='spotify-preview'
-            key={song.link}
-            thumbnailUrl={song.imageLink}
-            songName={song.name}
-            artistName={song.artist}
-            views={song.views + ' streams'}
+            key={song.songLink}
+            thumbnailUrl={song.thumbnailUrl}
+            songName={song.songName}
+            artistName={song.artistName}
+            artists={song.artists}
             duration={song.duration}
-            songLink={song.link}
+            songLink={song.songLink}
             platform={TYPE_SPOTIFY}
             handleAddClick={handleAddClick}
             playlistCreatorRef={playlistCreatorRef}
@@ -154,6 +189,7 @@ const SearchPage = () => {
       <div className='spotify-albums'>
         <h2 className='heading'>Spotify Albums</h2>
         <div className='spotify-album-list' style={{ display: 'flex', 'flex-wrap': 'wrap' }}>
+
           {spotifyAlbums.map((album) => (
             <PlaylistResult
               className={'spotify-album-preview'}
@@ -173,27 +209,11 @@ const SearchPage = () => {
               saveOnClick={saveOnClick}
             />
           ))}
-
-//         <div className='spotify-album-list' style={{display:'flex', 'flex-wrap': 'wrap'}}>
-
-//         {spotifyAlbums.map((album) => (
-//           <PlaylistResult
-//             className={'spotify-album-preview'}
-//             key={album.originSpotifyId}
-//             thumbnailUrl={album.coverImageURL}
-//             playlistName={album.name}
-//             artistName={album.author}
-//             views={album.popularity + ' views'}
-//             playlistLink={album.originSpotifyId}
-//           />
-//         ))}
-
         </div>
       </div>
 
       <div className='spotify-playlists'>
         <h2 className='heading'>Spotify Playlists</h2>
-
         <div className='spotify-playlist-list' style={{ display: 'flex', 'flex-wrap': 'wrap' }}>
           {spotifyPlaylists.map((playlist) => (
             <PlaylistResult
@@ -236,46 +256,11 @@ const SearchPage = () => {
             />
           ))}
 
-
-//         <div className='spotify-playlist-list' style={{display:'flex', 'flex-wrap': 'wrap'}}>
-//         {spotifyPlaylists.map((playlist) => (
-//           <PlaylistResult
-//             className={'spotify-playlist-preview'}
-//             key={playlist.originSpotifyId}
-//             thumbnailUrl={playlist.coverImageURL}
-//             playlistName={playlist.name}
-//             artistName={playlist.author}
-//             views={0 + ' views'}
-//             playlistLink={playlist.originSpotifyId}
-//           />
-//         ))}
-//           </div>
-//           </div>
-        
-//         <div className='youtube-videos'>
-//         <h2 className='heading'>Youtube Videos</h2>
-//         <div className='youtube-video-list' style={{display:'flex', 'flex-wrap': 'wrap'}}>
-//         {youtubeVideos.map((song) => (
-//           <SongResult
-//             className={'youtube-preview'}
-//             key={song.link}
-//             thumbnailUrl={song.imageLink}
-//             songName={song.name}
-//             artistName={song.artist}
-//             views={song.views + ' views'}
-//             duration={song.duration}
-//             songLink={song.link}
-//             platform='Youtube'
-//           />
-//         ))}
-        
-
         </div>
       </div>
 
       <div className='youtube-playlists'>
         <h2 className='heading'>Youtube Playlists</h2>
-
         <div className='youtube-playlist-list' style={{ display: 'flex', 'flex-wrap': 'wrap' }}>
           {spotifySongList.map((song) => (
             <PlaylistResult
@@ -309,25 +294,8 @@ const SearchPage = () => {
         <h2 className='heading'>Spotify Artists</h2>
         <div className='spotify-artist-list' style={{display:'flex', 'flex-wrap': 'wrap'}}>
           </div>
-
-        <div className='youtube-playlist-list' style={{display:'flex', 'flex-wrap': 'wrap'}}>
-        {youtubePlaylists.map((playlist) => (
-          <PlaylistResult
-            className={'youtube-playlist-preview'}
-            key={playlist.name}
-            thumbnailUrl={playlist.coverImageURL}
-            playlistName={playlist.name}
-            artistName={playlist.author}
-            songs={[]}
-            // views={song.views + ' views'}
-            // duration={song.duration}
-            playlistLink={playlist.originId}
-          />
-        ))}
-
           </div>
-          </div> 
-{/*
+
         <div className='unifi-playlists'>
         <h2 className='heading'>Uni.fi Playlists</h2>
         <div className='unifi-playlist-list' style={{display:'flex', 'flex-wrap': 'wrap'}}>
@@ -338,6 +306,59 @@ const SearchPage = () => {
     </div>
   );
 };
+
+const Preview = ({ title, author, views, platform }) => (
+  <div className='preview'>
+    <h4 className='preview-title'>{title}</h4>
+    <p className='preview-author'>Author: {author}</p>
+    <p className='preview-info'>Views/Streams: {views}</p>
+    <p className='preview-platform'>Platform: {platform}</p>
+  </div>
+);
+
+
+const generateRandomSongs = (count) => {
+  const songs = [];
+  for (let i = 0; i < count; i++) {
+    const song = {
+      id: i,
+      title: `Song ${i}`,
+      author: `Author ${i}`,
+      streams: Math.floor(Math.random() * 1000000),
+    };
+    songs.push(song);
+  }
+  return songs;
+};
+
+const generateRandomPlaylists = (count) => {
+  const playlists = [];
+  for (let i = 0; i < count; i++) {
+    const playlist = {
+      id: i,
+      title: `Playlist ${i}`,
+      author: `Author ${i}`,
+      streams: Math.floor(Math.random() * 1000000),
+    };
+    playlists.push(playlist);
+  }
+  return playlists;
+};
+
+const generateRandomVideos = (count) => {
+  const videos = [];
+  for (let i = 0; i < count; i++) {
+    const video = {
+      id: i,
+      title: `Video ${i}`,
+      author: `Author ${i}`,
+      views: Math.floor(Math.random() * 1000000),
+    };
+    videos.push(video);
+  }
+  return videos;
+};
+
 
 
 export default SearchPage;
