+++ server/app.js
@@ -148,8 +148,6 @@ setupCollections().catch(console.dir);
 // api
 var spotifyRouter = require('./routes/spotify');
 var youtubeRouter = require('./routes/youtube');
-var youtubeSearchRouter = require('./routes/ytSearch');
-var spotifySearchRouter = require('./routes/spSearch');
 
 var app = express();
 
@@ -165,8 +163,6 @@ app.use('/', indexRouter);
 // API
 app.use('/spotify', spotifyRouter);
 app.use('/youtube', youtubeRouter);
-app.use('/yt-search', youtubeSearchRouter);
-app.use('/sp-search', spotifySearchRouter);
 
 // login
 app.use('/login', loginRouter);
